Create test.file.md
test to create a folder in a repository:

Click on New file
On the text field for the file name, first write the folder name you want to create
Then type /. This creates a folder and add the file name. 

Always create a file because GitHub does not create and manage empty folders.